Load aliases and functions after other zsh files
'git' gets defined as an alias for 'hub' in the 'git' file. However, 'g' gets
defined as an alias for 'git' in the 'aliases' file. We need to ensure to load
the 'aliases' file after the 'git' one so that 'g' is an alias for 'hub' and not
for 'git'.