Load rbenv config before node config
We want to ensure that in the PATH, `node_modules/.bin` appears before
`~/.rbenv/shims` and `~/.rbenv/bin`. The specific reason why we need
this change is because of Jasmine. Jasmine has both an NPM package and
Ruby gem. In JavaScript projects in which Jasmine-the-package is used we
would like to prefer this over Jasmine-the-gem.