Remove oh-my-zsh, optimize startup time
Now that we've moved .zshrc to .zshenv, we've actually slowed down Vim,
because this means that every time Vim shells out it's having to do more
things. This is particularly noticeable when using vim-tmux-navigator to
switch from Vim to an adjacent tmux pane.

It turns out that Homebrew's `brew` executable is the main culprit here
-- every time it is run, it takes 0.15s. Every time we were shelling out
to it we were trying to get the path of an installed package -- while
this is useful, it's not worth it, so hardcode those paths everywhere
they are referred to.

This commit still slows down Vim a little bit, but there isn't anything
we can do about that right now. I might consider switching to fish to
see if that will help.