Make Ctrl-W act like Bash
Ctrl-W deletes the word before the cursor. What a word means differs
between zsh and Bash. In zsh, a word includes punctuation such as `/`.
This is annoying because oftentimes I just want to delete a part of a
path. This commit changes how zsh works so it uses Bash's definition a
word so that Ctrl-W stops at characters like `/`.