Show partial commands in status line
This will make it so that if you type e.g. "ciw", the "c" and "i" will
show up in the status bar, to let you know that you need to complete the
command. (I really wanted to show something more conspicuous, but it
doesn't seem there's a way to do that.)