Split zshenv into zshrc and zshenv again
Having one giant zshenv means that whenever Vim shells out, it costs a
lot of time. This is especially noticeable when using vim-tmux-runner to
run tests.

Now we have two files: zshenv and zshrc. zshenv contains the minimal
amount of code to start a shell (basically, setting the PATH and that's
it); zshrc contains the rest.

The reason why we smushed them together in the first place was so that
Vim could run executables installed by rbenv and NVM. Since rbenv and
NVM are version managers for Node and Ruby respectively, installed
packages will show up in whichever directory corresponds to the current
version, which could change in the future. So previously, zshenv loaded
rbenv and NVM into memory. This incurs a high startup cost, so instead
of doing that, we simply manually add the latest version of Node and
Ruby to the PATH -- but only under a non-login shell (otherwise we let
NVM and rbenv do its thing).