tmux: Attempt to fix issues with scrolling
In tmux 2.1 under iTerm2, scrolling with the mouse wheel (or a trackpad)
doesn't seem to do anything unless copy mode is first entered manually.
I found some stuff on Stack Overflow that seems to fix it pretty well,
except that now, scrolling is too fast. There's a tmux plugin that fixes
this for good, so that will come next.