Revert back to christoomey's VTR plugin
Using Neovim's async capabilities seems to be interfering with how VTR
works. VTR will send a sequence to the active tmux buffer to clear the
prompt before it runs a given command, but sometimes what happens is the
command runs before the clear sequence is run, and sometimes it runs
after. I don't know how to solve this right now, so let's just revert
and come back to this.