If 1st arg is dir, NERDTree opens at that dir
Before we switched to Neovim, if you opened Vim with a directory as the
first argument, NERDTree would open pointing to that directory. That now
no longer happens for whatever reason, and it seems to be that the
current working directory isn't set correctly. I think this may be a
Neovim thing, but in any case, explicitly set this behavior.