Use cwd for CtrlP's working path
By default CtrlP tries to use the currently open file's directory as the
working path. This is fine except that if you open Vim using a directory
as the first argument, then the currently open file is an empty buffer.
So CtrlP uses the directory that your shell was in when you launched
Vim. This commit fixes that.