Add two new mappings: ,el and ,ef
Usually when Neomake runs to lint the file you're working on, if there
are any errors, you do not want to see them in the quickfix window, as
this can be very annoying. However, if you are saving a file that you
haven't previously run through a linter and you have a lot of errors,
you may need to work through those errors and fix them one by one, so it
can be helpful to do two things: 1) show the quickfix window and 2) jump
all the way up to the first error.