Revise colorscheme toggling code
We have custom functions to toggle the colorscheme from dark to light.
We also set custom colors for the gutter and custom highlight groups.
Some of these colors do not get reset correctly when going from light
back to dark again.

To fix this, set these colors in the same functions that establish the
colorscheme.