Fix tmux + nvm so that we do not get an error
When starting tmux, we get this error from nvm:

    nvm is not compatible with the npm config "prefix" option: currently set to "/usr/local"
    Run `npm config delete prefix` or `nvm use --delete-prefix v8.8.1 --silent` to unset it.

This seems to be caused by an incorrect PATH. This happens because when
we set the PATH, we are prepending to it rather than resetting it
completely. Although tmux inherits its parent shell's PATH, it also
reloads the zsh configuration, so we end up with duplicate entries in
the PATH. Resetting the PATH each time zsh is initialized ensures that
PATH is the same regardless of whether we are inside tmux or outside it.