Fix highlighting of incremental search terms
For some reason, this got changed. I'm not sure if it was due to a Vim
update or not, but now whenever you're searching for something, the the
first match appears as orange with no background, whereas other matches
appear with a yellow background. So it's confusing because the first
match should pop out more than the other matches.