Replace matchit with matchup
Oftentimes after a while in Vim whenever I try to use `%` to bounce
between curly braces or somesuch, I get some kind of error like "can't
find matching bracket" or something that seems to be happening inside of
the matchit code. matchit has not been updated in forever, and matchup
is a newer package that seems to work a lot better.