Use nodenv instead of NVM
NVM is really annoying because it depends on the PATH to be set to a
certain thing. If you're trying to run a command instead via a Node
package through an editor like Vim, then you're out of luck — you're
using the default Node whether you like it or not.

But with nodenv, there's no PATH manipulation: you have the same set of
shims, and these shims dispatch to a certain version of that executable
that depends on whatever Node version nodenv is set to. This means that
you can tell your editor to use these shims, and they will always work.