Fix Ale overrides so they apply to filetype not extension
For instance, if Vim applies the file type of "javascript" to a file but
that file does not end in *.js, then it will not be run through
Prettier unless we change the extension. If we just use file type rather
than extension we don't have to baby these config files so much.