Support nested sessions in tmux
If you open a tmux session within an existing tmux session, there is
usually no way to interact with that child session because any
keystrokes you enter will be intercepted by the parent. This commit
makes it so that you can press F12 to change which session has control.
This is useful specifically when running `overmind` and connecting to an
existing tmux instance.