Easily switch btw light/dark modes in concert
When I am coding outside, I very often want to switch everything to
light mode so that I can see better — not only Vim, but also iTerm and
tmux. The problem is that this is a pain. This commit makes it a lot
easier!

To accomplish this, a few things were needed:

* iTerm's Python API so that we can automate between the Solarized Dark
  and Solarized Light profiles programmatically
* Shell scripting so that we can change color settings within tmux
  programmatically and also kick off the iTerm change
* Slight tweaks to the existing color configuration within Vim to kick
  off both iTerm and tmux changes

This was a lot of fun to figure out! I was very inspired by the
following two posts:

* <https://grrr.tech/posts/2020/switch-dark-mode-os/>
* <https://arslan.io/2021/02/15/automatic-dark-mode-for-terminal-applications/>