Fix PATH getting overridden in subshells
I started receiving this issue all of a sudden where when a command was
executed in a subshell inside of Vim (for instance, when using Ctrl-P),
PATH was getting overridden. It seems that environment variables are
somehow now getting inherited in non-interactive subshells so there is
no need to have a `~/.zshenv` which would presumably only get run in
such subshells.