Pin CoC and its plugins to avoid an incompat error
For some reason, there is an incompatibility between the latest versions
of `coc.nvim` and `coc-eslint`, where every time you update a file
you'll get an immediate error from somewhere within CoC. If you disable
the `eslint` plugin then the error will go away.

To prevent this from happening on a new computer in the future, pin CoC
and its plugins so that they'll never update.