Fix blowing away macOS's PATH
Currently we blow away PATH and then add our own additions to PATH.
This means that the initial value of PATH, which macOS sets via
/etc/zprofile, gets discarded. We also show a warning if /etc/zprofile,
asking the user to remove this file.

macOS's default PATH is probably useful in some way, however, and we
shouldn't be blowing it away. So we don't do that anymore. This does
create a slight problem in tmux where /etc/zprofile gets reloaded, and
so the PATH gets rearranged (the stuff that macOS adds appears first,
then our additions, instead of the other way around). We include a
little hack to prevent this from happening.