Add colors to tig
This is just based on [some dotfiles I found][1]. I will go and make
these consistent with Solarized later.

Note that we have to alias `tig` so that it uses a `TERM` of
`xterm-256colors` for now due to `tig` (or maybe `ncurses`, not sure)
not supporting our custom `tmux-256colors` file. I will also fix this
later (via [these][2] [sources][3])

[1]: https://github.com/cbertoldi/dotfiles/blob/master/tigrc.symlink
[2]: https://github.com/jonas/tig/issues/1210
[3]: https://superuser.com/questions/1420142/no-colors-in-weechat-tig-htop-ranger-etc