Fix workflow issues with Obsidian sync script
Sometimes, when the script to sync the Obsidian vault ends, it doesn't
leave the branch in a clean slate, and therefore, the next time it runs,
it will fail (because either the current branch is dirty, or is not
`main`, or both). Update the script to handle more states to reduce the
change that it fails to run.