Don't set up new Git repos with hooks
I added this a while back to make ctags always work in projects, but I
honestly don't use ctags for anything. And anyway, I don't like how
when a new Git repo is initialized, it uses symlinks for hooks. It means
that if you want to move your dotfiles somewhere else, suddenly you
break hooks in a bunch of repos.